Split the identifier validators into one file per country
