+++ src/redacters/local_rules/validators/africa/mod.rs
@@ -1,0 +1,5 @@
+//! Identifiers of Africa, one module per country.
+
+mod south_africa;
+
+pub use south_africa::*;
