+++ src/redacters/local_rules/validators/americas/mod.rs
@@ -1,0 +1,9 @@
+//! Identifiers of the Americas, one module per country.
+
+mod brazil;
+mod canada;
+mod us;
+
+pub use brazil::*;
+pub use canada::*;
+pub use us::*;
