+++ src/redacters/local_rules/validators/asia/mod.rs
@@ -1,0 +1,7 @@
+//! Identifiers of Asia, one module per country.
+
+mod china;
+mod india;
+
+pub use china::*;
+pub use india::*;
