+++ src/redacters/local_rules/validators/oceania/mod.rs
@@ -1,0 +1,5 @@
+//! Identifiers of Oceania, one module per country.
+
+mod australia;
+
+pub use australia::*;
